+++ api/src/team/application/routes.js
@@ -5,6 +5,7 @@ import { certificationCenterMembershipAdminRoutes } from './certification-center
 import { certificationCenterMembershipRoute } from './certification-center-membership/certification-center-membership.route.js';
 import { membershipAdminRoutes } from './membership/membership.admin.route.js';
 import { membershipRoutes } from './membership/membership.route.js';
+import { organizationMemberIdentitiesRoute } from './membership/organization-member-identities.route.js';
 import { organizationInvitationAdminRoutes } from './organization-invitations/organization-invitation.admin.route.js';
 import { organizationInvitationRoutes } from './organization-invitations/organization-invitation.route.js';
 import { prescriberInformationsRoute } from './prescriber-informations.route.js';
@@ -19,6 +20,7 @@ const register = async function (server) {
     ...certificationCenterMembershipAdminRoutes,
     ...membershipAdminRoutes,
     ...membershipRoutes,
+    ...organizationMemberIdentitiesRoute,
     ...userOrgaSettingsRoute,
     ...prescriberInformationsRoute,
     ...organizationInvitationRoutes,
