+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -401,57 +401,6 @@ describe('Acceptance | Application | organization-controller', function () {
     });
   });
 
-  describe('GET /api/organizations/{id}/member-identities', function () {
-    it('should return the members identities as JSON API', async function () {
-      // given
-      const organizationId = databaseBuilder.factory.buildOrganization().id;
-      const otherOrganizationId = databaseBuilder.factory.buildOrganization().id;
-      const member1 = databaseBuilder.factory.buildUser();
-      const member2 = databaseBuilder.factory.buildUser();
-      const otherOrganizationMember = databaseBuilder.factory.buildUser();
-
-      databaseBuilder.factory.buildMembership({ userId: member1.id, organizationId });
-      databaseBuilder.factory.buildMembership({ userId: member2.id, organizationId });
-      databaseBuilder.factory.buildMembership({
-        userId: otherOrganizationMember.id,
-        organizationId: otherOrganizationId,
-      });
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject({
-        method: 'GET',
-        url: `/api/organizations/${organizationId}/member-identities`,
-        headers: generateAuthenticatedUserRequestHeaders({ userId: member1.id }),
-      });
-
-      // then
-      const expectedResult = {
-        data: [
-          {
-            attributes: {
-              'first-name': member1.firstName,
-              'last-name': member1.lastName,
-            },
-            id: member1.id.toString(),
-            type: 'member-identities',
-          },
-          {
-            attributes: {
-              'first-name': member2.firstName,
-              'last-name': member2.lastName,
-            },
-            id: member2.id.toString(),
-            type: 'member-identities',
-          },
-        ],
-      };
-      expect(response.statusCode).to.equal(200);
-      expect(response.result).to.deep.equal(expectedResult);
-    });
-  });
-
   describe('PATCH /api/organizations/{id}/resend-invitation', function () {
     let clock;
     const now = new Date('2022-12-25');
